Move areCompatible
