+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/expression/predicate/operator/comparison/In.java
@@ -137,7 +137,7 @@ public class In extends EsqlScalarFunction {
         if (EsqlDataTypes.isSpatial(left) && EsqlDataTypes.isSpatial(right)) {
             return left == right;
         }
-        return EsqlDataTypes.areCompatible(left, right);
+        return DataType.areCompatible(left, right);
     }
 
     @Override
